ts merge (THX OLEG111) and controllers prepare + small fixes
