+++ .gitignore
@@ -1,11 +1,4 @@
-# Ignore node_modules directory
 node_modules/
 uploads/
-
-# Ignore .env files
 .env
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
+.*.local
