+++ backend/Dockerfile
@@ -1,8 +1,8 @@
 # Use an official Node.js runtime as a parent image
-FROM node:22-alpine
+FROM node:22-alpine3.19
 
 # Set the working directory in the container
-WORKDIR /usr/src/app
+WORKDIR /usr/app
 
 # Copy package.json and package-lock.json to the working directory
 COPY package*.json ./
@@ -17,4 +17,4 @@ COPY . .
 EXPOSE 6969
 
 # Define the command to run the app
-CMD ["node", "app.js"]
+CMD ["npm", "run" , "start"]
