+++ backend/models/Post.ts
@@ -16,4 +16,4 @@ const PostSchema = new Schema(
 
 const PostModel = model("Post", PostSchema);
 
-module.exports = PostModel;
+export default PostModel;
