+++ backend/models/User.ts
@@ -24,4 +24,4 @@ const UserSchema = new Schema(
 
 const UserModel = model("User", UserSchema);
 
-module.exports = UserModel;
+export default UserModel;
