+++ backend/package.json
@@ -5,7 +5,7 @@
   "main": "app.js",
   "type": "commonjs",
   "scripts": {
-    "start": "nodemon app.js"
+    "start": "nodemon app.ts"
   },
   "keywords": [],
   "author": "",
@@ -24,8 +24,16 @@
     "nodemon": "^2.0.20",
     "react-router-dom": "^6.23.1",
     "react-time-ago": "^7.3.3",
-    "sharp": "^0.33.5",
     "socket.io": "^4.7.5",
     "socket.io-client": "^4.7.5"
+  },
+  "devDependencies": {
+    "@types/bcryptjs": "^2.4.6",
+    "@types/cookie-parser": "^1.4.7",
+    "@types/express": "^4.17.21",
+    "@types/jsonwebtoken": "^9.0.6",
+    "@types/multer": "^1.4.12",
+    "@types/node": "^22.5.4",
+    "ts-node": "^10.9.2"
   }
 }
