+++ backend/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "outDir": "build",
+    "sourceMap": true,
+    "noImplicitAny": true,
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "lib": ["DOM", "es2017"],
+    "plugins": [{ "name": "typescript-plugin-css-modules" }],
+    "target": "ES2022",
+    "typeRoots": ["./node_modules/@types", "./types"]
+  },
+  "include": ["./**/*"]
+}
