+++ backend/types/types.ts
@@ -1,0 +1,12 @@
+import { Request } from "express";
+
+declare global {
+  namespace Express {
+    interface Request {
+      user: {
+        id: string;
+        username: string;
+      };
+    }
+  }
+}
