+++ docker-compose.yaml
@@ -1,30 +1,28 @@
 services:
 
+  mongo:
+    image: mongo
+    restart: always
+    # environment:
+      # MONGO_INITDB_ROOT_USERNAME: root
+      # MONGO_INITDB_ROOT_PASSWORD: example
+    ports:
+      - 27017:27017
+    
   frontend:
-    build:
+    build: 
       context: ./frontend
       dockerfile: Dockerfile
     ports:
       - 80:80
-      - 443:443
     depends_on:
       - backend
-    networks:
-      - custom-network 
-    volumes: 
-      - /etc/letsencrypt:/etc/letsencrypt:ro
-      - ./uploads:/usr/share/nginx/uploads
+  
   backend:
-    build:
+    build: 
       context: ./backend
       dockerfile: Dockerfile
     ports:
       - 6969:6969
-    networks:
-      - custom-network
-    volumes:
-      - ./uploads:/usr/src/uploads
-
-networks:
-  custom-network:
-    driver: bridge
+    depends_on:
+      - mongo
